# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pomodoro Api",
        "version": "0.1"
    },
    "paths": {
        "/api/pomodoros": {
            "get": {
                "tags": [
                    "pomodoros"
                ],
                "description": "return a list of pomodoros timers",
                "operationId": "a91850d394fead9e86bc2d3b8c0c88c6",
                "responses": {
                    "200": {
                        "description": "AOK"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "/api/pomodoros/{uuid}": {
            "get": {
                "tags": [
                    "pomodoros"
                ],
                "description": "return a pomodoros timer based on uuid",
                "operationId": "605f8d3b41091af9563c74ffd6d4c885",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Find pomodoro by uuid",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AOK"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate authentication token",
                "description": "Authenticate user and generate authentication token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        }
    }
}